+++ src/Ref.php
@@ -22,26 +22,15 @@ use const FILTER_VALIDATE_URL;
 
 final class Ref
 {
-    /** @var string */
-    public $title = '';
+    public string $title = '';
 
-    /** @var string */
-    public $type = '';
+    public string $type = '';
 
-    /** @var ?object */
-    public $json;
+    public ?object $json = null;
 
-    /**
-     * @var string
-     * @readonly
-     */
-    public $href = '';
+    public string $href = '';
 
-    /**
-     * @var ?object
-     * @readonly
-     */
-    public $schema;
+    public ?object $schema = null;
 
     public function __construct(string $ref, SplFileInfo $file, object $schema)
     {
@@ -62,7 +51,6 @@ final class Ref
         assert(isset($target->type));
         assert(is_string($target->type));
         $this->json = $target;
-        /** @psalm-suppress InaccessibleProperty */
         $this->type = $target->type;
         $title = $target->title ?? $path;
         assert(is_string($title));
@@ -72,11 +60,10 @@ final class Ref
     private function getExternalRef(string $ref, SplFileInfo $file): void
     {
         $filePath = $this->getFilePath($ref, $file);
-        $this->json = $schema = (object) json_decode((string) file_get_contents($filePath));
-        /** @psalm-suppress MixedAssignment */
-        $this->type = $schema->type ?? '';
-        /** @psalm-suppress MixedAssignment */
-        $this->title = $schema->title ?? '';
+        $this->json = (object) json_decode((string) file_get_contents($filePath));
+        $schema = (object) json_decode((string) file_get_contents($filePath));
+        $this->type = isset($schema->type) && is_string($schema->type) ? $schema->type : '';
+        $this->title = isset($schema->title) && is_string($schema->title) ? $schema->title : '';
     }
 
     private function getFilePath(string $ref, SplFileInfo $file): string