+++ src/TagLinks.php
@@ -12,16 +12,12 @@ use function sprintf;
 
 use const PHP_EOL;
 
-/**
- * @psalm-pure
- */
-final class TagLinks implements Stringable
+/** @psalm-pure */
+final readonly class TagLinks implements Stringable
 {
-    /**
-     * @param array<Link> $links
-     */
+    /** @param array<Link> $links */
     public function __construct(
-        private readonly array $links
+        private array $links
     ) {
     }